requirements.txt
openai
streamlit
python-dotenv
pandas
markdown
pyyaml
nbformat

pytest
